Terminados los artistas y las canciones, faltan las listas de reproducción

+++ src/main/java/com/salesianostriana/dam/trianafy/service/SongService.java
@@ -1,6 +1,8 @@
 package com.salesianostriana.dam.trianafy.service;
 
 
+import com.salesianostriana.dam.trianafy.exception.EmptySongListException;
+import com.salesianostriana.dam.trianafy.exception.SongNotFoundException;
 import com.salesianostriana.dam.trianafy.model.Artist;
 import com.salesianostriana.dam.trianafy.model.Song;
 import com.salesianostriana.dam.trianafy.repos.SongRepository;
@@ -17,20 +19,32 @@ public class SongService {
 
     private final SongRepository repository;
 
+    private final PlaylistService playlistService;
+
     public Song add(Song song) {
         return repository.save(song);
     }
 
-    public Optional<Song> findById(Long id) {
-        return repository.findById(id);
+    public Song findById(Long id) {
+        return repository.findById(id).orElseThrow(() -> new SongNotFoundException(id));
     }
 
     public List<Song> findAll() {
-        return repository.findAll();
+        List<Song> songs = repository.findAll();
+        if(songs.isEmpty()) {
+            throw new EmptySongListException();
+        }
+        return songs;
     }
 
-    public Song edit(Song song) {
-        return repository.save(song);
+    public Song edit(Long id, Song toEdit) {
+        return repository.findById(id).map(song -> {
+            song.setTitle(toEdit.getTitle());
+            song.setYear(toEdit.getYear());
+            song.setAlbum(toEdit.getAlbum());
+            song.setArtist(toEdit.getArtist());
+            return repository.save(song);
+        }).orElseThrow(() -> new SongNotFoundException(id));
     }
 
     public void delete(Song song) {
@@ -38,12 +52,24 @@ public class SongService {
     }
 
     public void deleteById(Long id) {
-        repository.deleteById(id);
+        if(repository.existsById(id)) {
+            Song toDelete = findById(id);
+            playlistService.findPlaylistBySong(toDelete).stream().forEach(pl -> {
+                while(pl.getSongs().contains(toDelete)) {
+                    pl.deleteSong(toDelete);
+                }
+            });
+            repository.deleteById(id);
+        }
     }
 
     public List<Song> findByArtist(Artist artist) {
         return repository.findByArtist(artist);
 
+    }
+
+    public List<Song> findByTitle(String title) {
+        return repository.findByTitleIgnoreCase(title);
     }
 
     public boolean existsById(Long id) {
